Do not count "hidden" comments against the comment limit.
Hidden comments are comments that github considers to be on "outdated
diffs". In the event that farcy has reached its self-imposed comment limit, it
would no longer make new comments making it impossible from the pull request
view to receive feedback on new issues. Because those comments are hidden on
page load, ignoring them in the comment count _should_ avoid the breaking the
pull request view (the reason why #35 was filed and subsequently resolved).

In the event we break pull request views again, we can use the information
about hidden comments to delete as many as needed.

Resolve #78.